+++ src/layman/layer/util_test.py
@@ -39,6 +39,7 @@ def test_fill_in_partial_info_statuses():
                         'updated_at': datetime.datetime(2021, 9, 9, 9, 39, 59, 167846,
                                                         tzinfo=tz.FixedOffsetTimezone(offset=0, name=None)),
                         'bounding_box': [1870322.81512642, 6281928.49798181, 1892002.82941466, 6304200.72172059],
+                        'native_crs': 'EPSG:3857',
                         'access_rights': {'read': ['lay3', 'EVERYONE'], 'write': ['lay3', 'EVERYONE']},
                         'file': {'path': 'layers/name_of_layer/input_file/name_of_layer.geojson',
                                  'file_type': settings.FILE_TYPE_VECTOR},
@@ -93,6 +94,7 @@ def test_fill_in_partial_info_statuses():
             1892002.82941466,
             6304200.72172059
         ],
+        'native_crs': 'EPSG:3857',
         'access_rights': {
             'read': [
                 'lay3',