+++ tests/dynamic_data/publications/__init__.py
@@ -27,6 +27,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),
@@ -43,6 +44,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),
@@ -123,6 +125,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -159,6 +162,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -200,6 +204,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -226,6 +231,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [-15695801.072582014, -7341864.739114417, 15699816.562538767, 11122367.192100529],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/ne_110m_admin_0_boundary lines land +ěščřžýáí/ne_110m_admin_0_boundary_lines_land ížě.shp',
                     'gdal_prefix': '/vsizip/',
@@ -252,6 +258,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1669480.0, 6580973.000000007, 1675351.9999999802, 6586999.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -276,6 +283,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                         '_file': {
                             'path': '/layman_data_test/workspaces/dynamic_test_workspace/layers/zipped_tif_tfw_rgba_opaque/input_file/zipped_tif_tfw_rgba_opaque.zip/small_layer.geojson'
                         },
@@ -306,6 +314,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [868376.0, 522128.0, 940583.0, 593255.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_colortable_nodata_opaque/sample_tif_colortable_nodata_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -330,6 +339,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [-15695801.072582014, -7341864.739114417, 15699816.562538767, 11122367.192100529],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/ne_110m_admin_0_boundary lines land +ěščřžýáí/ne_110m_admin_0_boundary_lines_land ížě.shp',
                     'gdal_prefix': '/vsizip/',
@@ -354,6 +364,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [868376.0, 522128.0, 940583.0, 593255.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_colortable_nodata_opaque/sample_tif_colortable_nodata_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -380,6 +391,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [868376.0, 522128.0, 940583.0, 593255.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_colortable_nodata_opaque/sample_tif_colortable_nodata_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -407,6 +419,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -432,6 +445,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [-15695801.072582014, -7341864.739114417, 15699816.562538767, 11122367.192100529],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/ne_110m_admin_0_boundary lines land +ěščřžýáí/ne_110m_admin_0_boundary_lines_land ížě.shp',
                     'gdal_prefix': '/vsizip/',
@@ -459,6 +473,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [-15695801.072582014, -7341864.739114417, 15699816.562538767, 11122367.192100529],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/ne_110m_admin_0_boundary lines land +ěščřžýáí/ne_110m_admin_0_boundary_lines_land ížě.shp',
                     'gdal_prefix': '/vsizip/',
@@ -484,6 +499,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -511,6 +527,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1669480.0, 6580973.000000007, 1675351.9999999802, 6586999.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -536,6 +553,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [868376.0, 522128.0, 940583.0, 593255.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_colortable_nodata_opaque/sample_tif_colortable_nodata_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -563,6 +581,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [868376.0, 522128.0, 940583.0, 593255.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_colortable_nodata_opaque/sample_tif_colortable_nodata_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -588,6 +607,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1669480.0, 6580973.000000007, 1675351.9999999802, 6586999.0],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque/sample_tif_tfw_rgba_opaque.tif',
                     'gdal_prefix': '/vsizip/',
@@ -613,6 +633,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [-15695801.072582014, -7341864.739114417, 15699816.562538767, 11122367.192100529],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/ne_110m_admin_0_boundary_lines_land.shp',
                     'gdal_prefix': '/vsizip/',
@@ -1005,6 +1026,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),
@@ -1034,6 +1056,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -1079,6 +1102,7 @@ PUBLICATIONS = {
                     'name': 'small_layer',
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),
@@ -1125,6 +1149,7 @@ PUBLICATIONS = {
                     'name': 'small_layer_with_id',
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer_with_id.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -1173,6 +1198,7 @@ PUBLICATIONS = {
                     'name': 'small_zip_layer',
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'zip/small_layer.geojson',
                     'gdal_prefix': '/vsizip/',
@@ -1198,6 +1224,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),
@@ -1244,6 +1271,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),
@@ -1292,6 +1320,7 @@ PUBLICATIONS = {
                 Action(publication.internal.correct_values_in_detail, {
                     'exp_publication_detail': {
                         'bounding_box': [1571204.369948366, 6268896.225570714, 1572590.854206196, 6269876.33561699],
+                        'native_crs': 'EPSG:3857',
                     },
                     'file_extension': 'geojson',
                     'publ_type_detail': ('vector', 'sld'),